+++ bin/actions/create_resource_request_action.js
@@ -2,16 +2,16 @@ const ReferenceStrategy = require(`../strategies/reference_strategy`)
 const PackageStrategy = require(`../strategies/package_strategy`)
 const TransitStrategy = require(`../strategies/transit_strategy`)
 
-module.exports = ({ reference, addendum }) => {
+module.exports = (reference) => {
   return new Promise((resolve, reject) => {
-    const specifier = ReferenceStrategy.referenceToSpecifier(reference, addendum)
-    const PackageTool = PackageStrategy.of(specifier)
-    const TransitTool = TransitStrategy.of(specifier)
+    const archiveRequest = ReferenceStrategy.referenceToArchiveRequest(reference)
+    const PackageTool = PackageStrategy.of(archiveRequest)
+    const TransitTool = TransitStrategy.of(archiveRequest)
 
     resolve({
       PackageTool,
       TransitTool,
-      specifier
+      archiveRequest
     })
   })
 }