+++ bin/actions/install_archive_action.js
@@ -1,9 +1,9 @@
 const promoteArchiveAction = require('./promote_archive_action')
 const downloadArchiveAction = require('./download_archive_action')
 
-const installArchiveAction = (args) => {
-  return downloadArchiveAction(args)
-    .then((archiveRequest) => {
+const installArchiveAction = (reference) => {
+  return downloadArchiveAction(reference)
+    .then(({ archiveRequest }) => {
       return promoteArchiveAction(archiveRequest)
     })
 }