+++ bin/commands/download.js
@@ -1,11 +1,11 @@
-const createArchiveRequestAction = require('../actions/create_archive_request_action')
+const createArchiveRequestAction = require('../actions/create_resource_request_action')
 const downloadArchiveAction = require('../actions/download_archive_action')
 const AuditLog = require('../support/audit_log')
 
 module.exports = {
   registerVorpalCommand: (vorpal, applicationConfiguration) => {
     return vorpal
-      .command(`download <reference> [addendum]`)
+      .command(`download <reference>`)
       .option('-a, --audit', `Inspect the tools selected for a reference`)
       .description(`Download an archive.`)
       .validate(function (args) {
@@ -13,17 +13,22 @@ module.exports = {
       })
       .action((args, done) => {
 
+        const { reference } = args
         // TODO download works on one archive at a time, try `all` for every package? or *
         // TODO evaluate how useful audit is and how it works with a full install
 
         if (args.options.audit) {
-          createArchiveRequestAction(args)
-            .then(({ specifier, TransitTool, PackageTool }) => {
+          return createArchiveRequestAction(reference)
+            .then(({
+              archiveRequest,
+              TransitTool,
+              PackageTool
+            }) => {
               vorpal.log(
                 AuditLog.variableValue({
-                  uri: specifier.uri,
-                  version: specifier.version,
-                  archive: specifier.archive,
+                  uri: archiveRequest.uri,
+                  version: archiveRequest.version,
+                  archive: archiveRequest.archive,
                   io: TransitTool.name,
                   package: PackageTool.name
                 })
@@ -32,15 +37,13 @@ module.exports = {
             .then(() => done())
         }
 
-        if (!args.options.audit) {
-          downloadArchiveAction(args, vorpal)
-            .catch(err => {
-              vorpal.log(err.toString())
-            })
-            .then(() => {
-              return done()
-            })
-        }
+        downloadArchiveAction(reference, vorpal)
+          .catch(err => {
+            vorpal.log(err.toString())
+          })
+          .then(() => {
+            return done()
+          })
       })
   }
 }