+++ bin/commands/install.js
@@ -1,4 +1,4 @@
-const createArchiveRequestAction = require('../actions/create_archive_request_action')
+const createArchiveRequestAction = require('../actions/create_resource_request_action')
 const installArchiveAction = require('../actions/install_archive_action')
 
 const AuditLog = require('../support/audit_log')
@@ -6,21 +6,27 @@ const AuditLog = require('../support/audit_log')
 module.exports = {
   registerVorpalCommand: (vorpal, applicationConfiguration) => {
     return vorpal
-      .command(`install <reference> [addendum]`)
+      .command(`install <reference>`)
       .option('-a, --audit', `Inspect the tools selected for a reference`)
       .description(`Install an archive.`)
       .validate(function (args) {
         return true
       })
       .action((args, done) => {
+        const { reference } = args
+
         if (args.options.audit) {
-          createArchiveRequestAction(args)
-            .then(({ specifier, TransitTool, PackageTool }) => {
+          return createArchiveRequestAction(reference)
+            .then(({
+              archiveRequest,
+              TransitTool,
+              PackageTool
+            }) => {
               vorpal.log(
                 AuditLog.variableValue({
-                  uri: specifier.uri,
-                  version: specifier.version,
-                  archive: specifier.archive,
+                  uri: archiveRequest.uri,
+                  version: archiveRequest.version,
+                  archive: archiveRequest.archive,
                   io: TransitTool.name,
                   package: PackageTool.name
                 })
@@ -29,15 +35,15 @@ module.exports = {
             .then(() => done())
         }
 
-        if (!args.options.audit) {
-          installArchiveAction(args, vorpal)
-            .catch(err => {
-              vorpal.log(err.toString())
-            })
-            .then(() => {
-              return done()
-            })
-        }
+        installArchiveAction(reference, vorpal)
+          .catch(err => {
+            vorpal.log(err.toString())
+          })
+          .then((repoManifest) => {
+            // TODO get list of dependencies and call installArchiveAction on each
+            console.log(repoManifest)
+            return done()
+          })
       })
   }
 }