+++ bin/support/file_system.js
@@ -21,7 +21,9 @@ class FileSystem {
       }
     }
 
-    return fs.copy(from, to, copyOptions)
+    return new Promise((resolve) => {
+      resolve(fs.copy(from, to, copyOptions))
+    })
   }
 
   /**
@@ -36,15 +38,25 @@ class FileSystem {
     }
   }
 
-  static makeDirectory (directoryName) {
+  static createDirectory (directoryName) {
     directoryName = path.normalize(directoryName).split(path.sep)
     directoryName.forEach((sdir, index) => {
       const pathInQuestion = directoryName.slice(0, index + 1).join(path.sep)
 
-      if (!this.isDirectory(pathInQuestion) && pathInQuestion) {
+      if (pathInQuestion && !this.isDirectory(pathInQuestion)) {
         fs.mkdirSync(pathInQuestion)
       }
     })
+
+    return this
+  }
+
+  static removeDirectory (directoryName) {
+    directoryName = path.normalize(directoryName)
+
+    if (directoryName && this.isDirectory(directoryName)) {
+      fs.removeSync(directoryName)
+    }
 
     return this
   }
@@ -53,7 +65,7 @@ class FileSystem {
     return new Promise((resolve, reject) => {
       const { folder } = this.explodePath(path)
 
-      this.makeDirectory(folder)
+      this.createDirectory(folder)
 
       const writer = fs.createWriteStream(this.readPath(path), streamOptions)
 
@@ -72,7 +84,7 @@ class FileSystem {
 
   static read (path, streamOptions = { encoding: `binary` }) {
     return new Promise((resolve, reject) => {
-      fs.readFile(this.readPath(path), `binary`, (err, data) => {
+      fs.readFile(this.readPath(path), streamOptions, (err, data) => {
         if (err) {
           return reject(new Error(err))
         }