+++ bin/vaults/local_vault.js
@@ -1,21 +1,14 @@
 
-
 class LocalVault {
-  static getVersions(archiveName) {
+  static getVersions (archiveRequest) {
     return new Promise((resolve, reject) => {
-      resolve([specifier.version])
+      resolve([archiveRequest.version])
     })
   }
 
-  static matchAppropriateVersion(archiveName, versionRequest) {
+  static assignAppropriateVersion (archiveRequest) {
     return new Promise((resolve, reject) => {
-      resolve([specifier.version])
-    })
-  }
-
-  static assignAppropriateVersion(specifier) {
-    return new Promise((resolve, reject) => {
-      resolve(specifier)
+      resolve(archiveRequest)
     })
   }
 }