+++ test/strategies/reference_strategy_spec.js
@@ -5,44 +5,21 @@ const ReferenceStrategy = require('../../bin/strategies/reference_strategy')
 const applicationConfiguration = require(`../../bin/configurations/application`)
 
 test.cb('strategy: reference (folder)', t => {
-  t.plan(5)
+  t.plan(4)
 
   const {
     stagingPath,
-    addendum,
     version,
     archive,
     uri
-  } = ReferenceStrategy.referenceToSpecifier(`../folder/`, `/sub_folder/`)
+  } = ReferenceStrategy.referenceToArchiveRequest(`../folder/sub_folder/`)
 
-  t.is(addendum, `sub_folder/`, `addendum strips the first folder slash, if it exists`)
   t.is(version, `master`, `version defaults to master`)
   t.is(archive, `sub_folder`, `the archive is the last folder name`)
-  t.is(uri, `../folder/`, `uri is the base path to the resource and needs the addendum to resolve the full path`)
+  t.is(uri, `../folder/sub_folder/`, `uri is the base path to the resource`)
 
   t.is(stagingPath, `${applicationConfiguration.get('paths.staging')}/${archive}/`, `staging path is the archive name in staging folder`)
 
-  t.end()
-})
-
-test.cb('strategy: scope-to-resource (git)', t => {
-  t.plan(1)
-
-  applicationConfiguration.override({
-    'sources': {
-      'git_source': {
-        'pattern': 'source/group/resource',
-        'template': 'https://<%= username %>@repo.advisory.com/scm/eabui/<%= group %>.git#<%= version %> <%=resource %>.zip',
-        'constants': {
-          'username': 'jaworskm'
-        }
-      }
-    }
-  })
-
-  const uri = ReferenceStrategy.scopeToResource(`$git_source/def/ghi@1.2.3`)
-
-  t.is(uri, `https://jaworskm@repo.advisory.com/scm/eabui/def.git#1.2.3 ghi.zip`)
   t.end()
 })