+++ changelog.md
@@ -1,5 +1,13 @@
 # Bauble Changelog
 
+
+
+## 0.3.0
+
+- ADDED semver rules for install
+- ADDED manifest recursive call
+- REMOVED git support
+
 ## 0.2.2
 
 - ADDED load configuration from application (default), home, and project